+++ packages/components/fields/async-creatable-select-field/package.json
@@ -22,6 +22,7 @@
     "@commercetools-uikit/design-system": "10.18.4",
     "@commercetools-uikit/field-errors": "10.22.1",
     "@commercetools-uikit/field-label": "10.22.1",
+    "@commercetools-uikit/icons": "10.22.0",
     "@commercetools-uikit/spacings": "10.18.5",
     "@commercetools-uikit/utils": "10.21.0",
     "@emotion/core": "10.0.28",
