+++ packages/components/fields/async-creatable-select-field/src/async-creatable-select-field.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { render, fireEvent, waitFor } from '../../../../test/test-utils';
+import { render, fireEvent, waitFor } from '../../../../../test/test-utils';
 import AsyncCreatableSelectField from './async-creatable-select-field';
 
 // This component is used to enable easy testing.
