+++ packages/components/fields/async-creatable-select-field/src/async-creatable-select-field.visualroute.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 import { AsyncCreatableSelectField } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 const loadOptions = () =>
   Promise.resolve([
