+++ packages/components/fields/async-creatable-select-field/src/index.js
@@ -1,1 +1,3 @@
 export { default } from './async-creatable-select-field';
+
+export { default as version } from './version';
